requirements.txt
pyautogui
pyperclip
python-dotenv
google-generativeai
